Use integer and double types for lineage statistics sums

+++ source/EngineInterface/StatisticsEntry.h
@@ -19,11 +19,11 @@ struct LineageStatisticsEntry
     uint32_t colorBitset = 0;
     uint32_t numCreatures = 0;
     uint32_t numGenomes = 0;
-    float sumCreatureCells = 0;
-    float sumCreatureGenerations = 0;
-    float sumGenomeNodes = 0;
-    float sumMutationRates = 0;
-    float sumCreatureEnergy = 0;
+    uint64_t sumCreatureCells = 0;
+    uint64_t sumCreatureGenerations = 0;
+    uint64_t sumGenomeNodes = 0;
+    double sumMutationRates = 0;
+    double sumCreatureEnergy = 0;
     uint64_t representativeCellId = 0;  // Cell of a creature with the highest generation; 0 = not set
 
     uint64_t numCreatedCreatures = 0;  // Accumulated, never reset
